% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_expect_equal}
\alias{construct_expect_equal}
\title{Construct an \code{expect_equal} expression}
\usage{
construct_expect_equal(text_expr)
}
\arguments{
\item{text_expr}{A character vector of lines that, when executed produce a
single output.}
}
\value{
A character vector. The lines of text containing the
\code{expect_equal} code corresponding to the input, which will help to
write the test file based on an example detailed with roxgen. Remember that
this is something that you're intended to fill the gaps in later.
}
\description{
Construct an \code{expect_equal} expression from a character vector
containing an expression to be evaluated.
}
\examples{
text_expr <- c("sum(1, ", "2)")
cat(paste(text_expr, collapse = "\\n"))
construct_expect_equal(text_expr)
cat(paste(construct_expect_equal(text_expr), collapse = "\\n"))
}
