% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exemplar.R
\name{make_tests_shells_file}
\alias{make_tests_shells_file}
\title{Create the shell of a test file based on roxygen examples.}
\usage{
make_tests_shells_file(r_file_name, proj_dir = ".")
}
\arguments{
\item{r_file_name}{The name of the .R file within R/. Don't specify this as
"R/x.R", just use "x.R" for whichever file x it is. You can also omit the
.R for convenience, however using the wrong case (e.g. .r) will produce an
error.}

\item{proj_dir}{The directory of the R project for this package (defaults to
current directory). Note that this is the parent directory of man/.}
}
\value{
The shell of the test file is written into tests/testthat. It has the
same name as the .R file it was created from except it has "test_" tacked
onto the front.
}
\description{
Based on the roxygen-specified examples in a .R file in the R/ directory of
the package, create the shell of a test file, to be completed by the user.
}
