% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{text_eval_error}
\alias{text_eval_error}
\title{Does evaluation of text give an error?}
\usage{
text_eval_error(text_expr)
}
\arguments{
\item{text_expr}{The expression to be evaluated, as a character vector.}
}
\value{
\code{TRUE} if the code gives an error and \code{FALSE} otherwise.
}
\description{
Can a character vector (where each line is treated as a line of R code) be
evaluated as an R expression (or several R expressions) without giving an
error?
}
\examples{
text_eval_error("a <- 1")
text_eval_error("a <- ")
}
