% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TextEval}
\alias{TextEval}
\title{Evaluate a text string}
\usage{
TextEval(string)
}
\arguments{
\item{string}{The string to evaluate (as if it were a command).}
}
\description{
Evaluate a text string
}
\examples{
TextEval("3 + 4")
to.be.evaluated <- "var(c(1, 6, 8))"
TextEval(to.be.evaluated)

}
